Clean code from GeoContainerViewer2 and GeoViewer
