+++ app/views/advanced_searches/_map.html.erb
@@ -1,4 +1,4 @@
-<%= react_component("GeoContainerViewer2",
+<%= react_component("GeoContainerViewer",
       id: "results_map_container",
       props: {
         features: results.items_as_geojson.to_json,
