+++ app/views/containers/_map.html.erb
@@ -1,4 +1,4 @@
-<%= react_component('GeoContainerViewer2', props: {
+<%= react_component('GeoContainerViewer', props: {
     features: container.geojson,
     layers: container.geo_layers,
     mapHeight: container.map_height,
