+++ app/views/fields/_geometries.html.erb
@@ -1,2 +1,2 @@
-<%= react_component('GeoContainerViewer2', props: { features: features.to_json, layers: layers, catalog: @catalog.slug, maxBBZoom: 13,
+<%= react_component('GeoContainerViewer', props: { features: features.to_json, layers: layers, catalog: @catalog.slug, maxBBZoom: 13,
 locale: I18n.locale }) %>
